# Conflicts:
#	web/src/main/resources/META-INF/resources/WEB-INF/views/inspectionReport/js/inspectionReport.js

+++ service/src/main/java/com/nicico/sales/service/invoice/foreign/ForeignInvoiceService.java
@@ -1,21 +1,38 @@
 package com.nicico.sales.service.invoice.foreign;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ghasemkiani.util.icu.PersianCalendar;
 import com.ibm.icu.util.Calendar;
 import com.nicico.sales.annotation.Action;
 import com.nicico.sales.dto.InvoiceTypeDTO;
+import com.nicico.sales.dto.MaterialElementDTO;
+import com.nicico.sales.dto.UnitDTO;
 import com.nicico.sales.dto.contract.ContractDTO;
+import com.nicico.sales.dto.contract.ContractDetailDTO;
+import com.nicico.sales.dto.contract.ContractDiscountDTO;
+import com.nicico.sales.dto.contract.IncotermDTO;
 import com.nicico.sales.dto.invoice.foreign.*;
 import com.nicico.sales.enumeration.ActionType;
+import com.nicico.sales.enumeration.EContractDetailTypeCode;
+import com.nicico.sales.enumeration.EContractDetailValueKey;
 import com.nicico.sales.enumeration.ErrorType;
 import com.nicico.sales.exception.NotFoundException;
 import com.nicico.sales.exception.SalesException2;
+import com.nicico.sales.iservice.IContractDetailValueService2;
+import com.nicico.sales.iservice.contract.IContractDetailService;
 import com.nicico.sales.iservice.invoice.foreign.IForeignInvoiceService;
+import com.nicico.sales.model.entities.base.Unit;
+import com.nicico.sales.model.entities.contract.Contract;
+import com.nicico.sales.model.entities.contract.ContractDiscount;
 import com.nicico.sales.model.entities.invoice.foreign.ForeignInvoice;
 import com.nicico.sales.model.entities.invoice.foreign.ForeignInvoiceBillOfLading;
 import com.nicico.sales.model.entities.invoice.foreign.ForeignInvoiceItem;
+import com.nicico.sales.model.entities.warehouse.MaterialElement;
+import com.nicico.sales.repository.UnitDAO;
+import com.nicico.sales.repository.contract.ContractDAO;
 import com.nicico.sales.repository.invoice.foreign.ForeignInvoiceBillOfLadingDAO;
 import com.nicico.sales.repository.invoice.foreign.ForeignInvoiceDAO;
+import com.nicico.sales.repository.warehouse.MaterialElementDAO;
 import com.nicico.sales.service.GenericService;
 import com.nicico.sales.service.InvoiceTypeService;
 import com.nicico.sales.service.contract.ContractService;
@@ -27,23 +44,33 @@ import org.springframework.context.support.ResourceBundleMessageSource;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class ForeignInvoiceService extends GenericService<ForeignInvoice, Long, ForeignInvoiceDTO.Create, ForeignInvoiceDTO.Info, ForeignInvoiceDTO.Update, ForeignInvoiceDTO.Delete> implements IForeignInvoiceService {
 
+    private final UnitDAO unitDAO;
+    private final ContractDAO contractDAO;
+    private final ObjectMapper objectMapper;
     private final ContractService contractService;
     private final InvoiceTypeService invoiceTypeService;
     private final InvoiceNoGenerator invoiceNoGenerator;
+    private final MaterialElementDAO materialElementDAO;
     private final ResourceBundleMessageSource messageSource;
+    private final IContractDetailService contractDetailService;
     private final ForeignInvoiceItemService foreignInvoiceItemService;
+    private final IContractDetailValueService2 contractDetailValueService2;
     private final ForeignInvoicePaymentService foreignInvoicePaymentService;
-    private final ForeignInvoiceItemDetailService foreignInvoiceItemDetailService;
     private final ForeignInvoiceBillOfLadingDAO foreignInvoiceBillOfLadingDAO;
+    private final ForeignInvoiceItemDetailService foreignInvoiceItemDetailService;
     private final ForeignInvoiceBillOfLadingService foreignInvoiceBillOfLadingService;
+
 
     @Override
     @Transactional
@@ -67,6 +94,91 @@ public class ForeignInvoiceService extends GenericService<ForeignInvoice, Long, 
 
     @Override
     @Transactional
+    @Action(value = ActionType.Get)
+    public ContractDetailDataDTO.Info getContractDetailData(Long contractId) {
+
+        ContractDetailDataDTO.Info contractDetailData = new ContractDetailDataDTO.Info();
+
+        Contract contract = contractDAO.findById(contractId).orElseThrow(() -> new NotFoundException(Contract.class));
+        Long materialId = contract.getMaterialId();
+
+        // MOAS
+        Map<String, List<Object>> operationalDataOfMOASArticle = contractDetailValueService2.get(contractId, EContractDetailTypeCode.QuotationalPeriod, EContractDetailValueKey.MOAS, true);
+        if (operationalDataOfMOASArticle.size() != 0) {
+            List<Map<String, Object>> moas = (List<Map<String, Object>>) operationalDataOfMOASArticle.get(EContractDetailValueKey.MOAS.getId()).get(0);
+            List<ContractDetailDataDTO.MOASData> moasData = new ArrayList<>();
+            if (moas != null) moas.stream().forEach(moasItem -> {
+                Long materialElementId = Long.valueOf(moasItem.get("materialElement").toString());
+                MaterialElement materialElement = materialElementDAO.findById(materialElementId).orElseThrow(() -> new NotFoundException(MaterialElement.class));
+                moasItem.put("materialElement", modelMapper.map(materialElement, MaterialElementDTO.Info.class));
+                moasData.add(objectMapper.convertValue(moasItem, ContractDetailDataDTO.MOASData.class));
+            });
+            contractDetailData.setMOAS(moasData);
+        } else
+            throw new SalesException2(ErrorType.NotFound, "MOAS", "Contract QuotationalPeriod Article Not Found");
+
+        // TC
+        Map<String, List<Object>> tcs = contractDetailValueService2.get(contractId, EContractDetailTypeCode.Deduction, EContractDetailValueKey.TC, true);
+        if (tcs != null && tcs.size() != 0)
+            contractDetailData.setTc(new BigDecimal(tcs.get(EContractDetailValueKey.TC.name()).get(0).toString()));
+        else if (materialId == 3)
+            throw new SalesException2(ErrorType.NotFound, "rc", "Contract TC Article Not Found");
+        else
+            contractDetailData.setTc(null);
+
+        // RC
+        Map<String, List<Object>> operationalDataOfRCArticle = contractDetailValueService2.get(contractId, EContractDetailTypeCode.Deduction, EContractDetailValueKey.RC, true);
+        if (operationalDataOfRCArticle.size() != 0) {
+            List<Map<String, Object>> rcs = (List<Map<String, Object>>) operationalDataOfRCArticle.get(EContractDetailValueKey.RC.getId()).get(0);
+            List<ContractDetailDataDTO.RCData> rcData = new ArrayList<>();
+            if (rcs != null) rcs.stream().forEach(rcItem -> {
+                Long financeUnitId = Long.valueOf(rcItem.get("financeUnit").toString());
+                Long weightUnitId = Long.valueOf(rcItem.get("weightUnit").toString());
+                Long materialElementId = Long.valueOf(rcItem.get("materialElement").toString());
+                Unit financeUnit = unitDAO.findById(financeUnitId).orElseThrow(() -> new NotFoundException(UnitDTO.class));
+                Unit weightUnit = unitDAO.findById(weightUnitId).orElseThrow(() -> new NotFoundException(UnitDTO.class));
+                MaterialElement materialElement = materialElementDAO.findById(materialElementId).orElseThrow(() -> new NotFoundException(MaterialElement.class));
+                rcItem.put("financeUnit", modelMapper.map(financeUnit, UnitDTO.Info.class));
+                rcItem.put("weightUnit", modelMapper.map(weightUnit, UnitDTO.Info.class));
+                rcItem.put("materialElement", modelMapper.map(materialElement, MaterialElementDTO.Info.class));
+                rcData.add(objectMapper.convertValue(rcItem, ContractDetailDataDTO.RCData.class));
+            });
+            contractDetailData.setRc(rcData);
+        }
+        else if (materialId == 3)
+            throw new SalesException2(ErrorType.NotFound, "rc", "Contract RC Article Not Found");
+        else
+            contractDetailData.setRc(null);
+
+        // INCOTERM
+        Map<String, List<Object>> deliveryTerms = contractDetailValueService2.get(contractId, EContractDetailTypeCode.DeliveryTerms, EContractDetailValueKey.INCOTERM, true);
+        if (deliveryTerms != null && deliveryTerms.size() != 0)
+            contractDetailData.setIncoterm(modelMapper.map(deliveryTerms.get(EContractDetailValueKey.INCOTERM.getId()).get(0), IncotermDTO.Info.class));
+        else
+            throw new SalesException2(ErrorType.NotFound, "deliveryTerm", "Contract DeliveryTerms Article Not Found");
+
+        // Discounts
+        Map<String, List<Object>> discounts = contractDetailValueService2.get(contractId, EContractDetailTypeCode.Price, EContractDetailValueKey.DISCOUNT, true);
+        if (discounts != null && discounts.size() != 0)
+            contractDetailData.setDiscount(modelMapper.map(discounts.get(EContractDetailValueKey.DISCOUNT.getId()).get(0), ContractDiscountDTO.Info.class));
+
+        // Price Article Content
+        ContractDetailDTO.Info priceDetail = contractDetailService.getContractDetailByContractDetailTypeCode(contractId, materialId, EContractDetailTypeCode.Price);
+        if (priceDetail != null) {
+            contractDetailData.setPriceContent(priceDetail.getContent());
+        }
+
+        // QuotationalPeriod Article Content
+        ContractDetailDTO.Info quotationalPeriodDetail = contractDetailService.getContractDetailByContractDetailTypeCode(contractId, materialId, EContractDetailTypeCode.QuotationalPeriod);
+        if (quotationalPeriodDetail != null) {
+            contractDetailData.setQuotationalPeriodContent(quotationalPeriodDetail.getContent());
+        }
+
+        return contractDetailData;
+    }
+
+    @Override
+    @Transactional
     @Action(ActionType.Create)
     public ForeignInvoiceDTO.Info create(ForeignInvoiceDTO.Create request) {
 
@@ -82,12 +194,14 @@ public class ForeignInvoiceService extends GenericService<ForeignInvoice, Long, 
             ForeignInvoiceItemDTO.Create foreignInvoiceItemCreate = modelMapper.map(item, ForeignInvoiceItemDTO.Create.class);
             foreignInvoiceItemCreate.setForeignInvoiceId(foreignInvoice.getId());
             ForeignInvoiceItemDTO.Info foreignInvoiceItem = foreignInvoiceItemService.create(foreignInvoiceItemCreate);
-            item.getForeignInvoiceItemDetails().forEach(detail -> {
+            if (item.getForeignInvoiceItemDetails() != null) {
+                item.getForeignInvoiceItemDetails().forEach(detail -> {
 
-                ForeignInvoiceItemDetailDTO.Create foreignInvoiceItemDetailCreate = modelMapper.map(detail, ForeignInvoiceItemDetailDTO.Create.class);
-                foreignInvoiceItemDetailCreate.setForeignInvoiceItemId(foreignInvoiceItem.getId());
-                foreignInvoiceItemDetailService.create(foreignInvoiceItemDetailCreate);
-            });
+                    ForeignInvoiceItemDetailDTO.Create foreignInvoiceItemDetailCreate = modelMapper.map(detail, ForeignInvoiceItemDetailDTO.Create.class);
+                    foreignInvoiceItemDetailCreate.setForeignInvoiceItemId(foreignInvoiceItem.getId());
+                    foreignInvoiceItemDetailService.create(foreignInvoiceItemDetailCreate);
+                });
+            }
         });
 
         request.getForeignInvoicePayments().forEach(item -> item.setForeignInvoiceId(foreignInvoice.getId()));
@@ -114,9 +228,11 @@ public class ForeignInvoiceService extends GenericService<ForeignInvoice, Long, 
         int itemSize = foreignInvoice.getForeignInvoiceItems().size();
         for (int i = 0; i < itemSize; i++) {
             ForeignInvoiceItem foreignInvoiceItem = foreignInvoice.getForeignInvoiceItems().get(i);
-            int itemDetailSize = foreignInvoiceItem.getForeignInvoiceItemDetails().size();
-            for (int j=0; j<itemDetailSize; j++) {
-                foreignInvoiceItemDetailService.delete(foreignInvoiceItem.getForeignInvoiceItemDetails().get(j).getId());
+            if (foreignInvoiceItem.getForeignInvoiceItemDetails() != null) {
+                int itemDetailSize = foreignInvoiceItem.getForeignInvoiceItemDetails().size();
+                for (int j = 0; j < itemDetailSize; j++) {
+                    foreignInvoiceItemDetailService.delete(foreignInvoiceItem.getForeignInvoiceItemDetails().get(j).getId());
+                }
             }
             foreignInvoiceItemService.delete(foreignInvoice.getForeignInvoiceItems().get(i).getId());
         }
@@ -124,9 +240,9 @@ public class ForeignInvoiceService extends GenericService<ForeignInvoice, Long, 
         foreignInvoice.getForeignInvoicePayments().forEach(item -> foreignInvoicePaymentService.delete(item.getId()));
 
         int billLadingSize = request.getBillLadingIds().size();
-        for (int i=0; i<billLadingSize; i++) {
+        for (int i = 0; i < billLadingSize; i++) {
             List<ForeignInvoiceBillOfLading> allByBillOfLandingIdAAndForeignInvoiceId = foreignInvoiceBillOfLadingDAO.findAllByBillOfLandingIdAndForeignInvoiceId(request.getBillLadingIds().get(i), foreignInvoice.getId());
-            for (int j=0; j<allByBillOfLandingIdAAndForeignInvoiceId.size(); j++) {
+            for (int j = 0; j < allByBillOfLandingIdAAndForeignInvoiceId.size(); j++) {
                 foreignInvoiceBillOfLadingService.delete(allByBillOfLandingIdAAndForeignInvoiceId.get(j).getId());
             }
         }
@@ -138,11 +254,13 @@ public class ForeignInvoiceService extends GenericService<ForeignInvoice, Long, 
             ForeignInvoiceItemDTO.Create foreignInvoiceItemCreate = modelMapper.map(item, ForeignInvoiceItemDTO.Create.class);
             foreignInvoiceItemCreate.setForeignInvoiceId(foreignInvoice.getId());
             ForeignInvoiceItemDTO.Info foreignInvoiceItem = foreignInvoiceItemService.create(foreignInvoiceItemCreate);
-            item.getForeignInvoiceItemDetails().forEach(detail -> {
-                ForeignInvoiceItemDetailDTO.Create foreignInvoiceItemDetailCreate = modelMapper.map(detail, ForeignInvoiceItemDetailDTO.Create.class);
-                foreignInvoiceItemDetailCreate.setForeignInvoiceItemId(foreignInvoiceItem.getId());
-                foreignInvoiceItemDetailService.create(foreignInvoiceItemDetailCreate);
-            });
+            if (item.getForeignInvoiceItemDetails() != null) {
+                item.getForeignInvoiceItemDetails().forEach(detail -> {
+                    ForeignInvoiceItemDetailDTO.Create foreignInvoiceItemDetailCreate = modelMapper.map(detail, ForeignInvoiceItemDetailDTO.Create.class);
+                    foreignInvoiceItemDetailCreate.setForeignInvoiceItemId(foreignInvoiceItem.getId());
+                    foreignInvoiceItemDetailService.create(foreignInvoiceItemDetailCreate);
+                });
+            }
         });
 
         request.getForeignInvoicePayments().forEach(item -> item.setForeignInvoiceId(foreignInvoice.getId()));
@@ -159,15 +277,4 @@ public class ForeignInvoiceService extends GenericService<ForeignInvoice, Long, 
         return foreignInvoiceDTO;
     }
 
-    @Override
-    @Transactional
-    @Action(ActionType.Delete)
-    public void delete(Long aLong) {
-    }
-
-    @Override
-    @Transactional
-    @Action(ActionType.DeleteAll)
-    public void deleteAll(ForeignInvoiceDTO.Delete request) {
-    }
 }
